Task #712 Stage 2-3 (GREEN): u32 vert_offset 게이트 signed 정정
Root cause:
  HwpUnit = u32 (src/model/mod.rs:21) 라서 음수 vertical_offset
  (예: -1796 HU = 0xFFFFF8FC = 4294965500u32) 이 unsigned 양수로
  저장됨. `vertical_offset > 0` u32 비교 게이트가 음수 비트표현도
  통과시키고, 후속 `as i32` 캐스트에서 음수가 그대로 적용 →
  PartialTable 표가 음수만큼 위로 점프, 직전 inline TAC 표 영역
  침범 (~12-16 px).

  비-Partial 경로 (table_layout.rs:1069+) 는 동일 분기에
  `pushed = raw_y.max(y_start)` 클램프가 있어 음수 무력화. Partial
  경로에는 클램프가 없어 본 결함 표면화.

정정:
  - src/renderer/layout/table_partial.rs:62-71 — gate 를 signed 비교
    `(vertical_offset as i32) > 0` 로 변경. 양수만 적용, 0/음수는
    y_start 유지 (비-Partial 경로 효과와 동등).
  - src/renderer/layout.rs:2673-2685 — pt_y_start 게이트 동일 정정.

검증:
  - tests/issue_712.rs PASS (pi586=[148.88..1028.25], 침범 0)
  - cargo test --release: 1252 passed, 0 failed, 5 ignored (회귀 0)
  - 페이지 37 (분할 표 연결) 회귀 확인 — !is_continuation 가드로 분할
    연결 분기는 무영향.

Stage 2 분석 보고서 + Stage 3 GREEN 보고서 동봉.

+++ src/renderer/layout.rs
@@ -2695,14 +2695,16 @@ impl LayoutEngine {
         let pt_mt = measured_tables.iter().find(|mt|
             mt.para_index == para_index && mt.control_index == control_index
         );
-        // 비-TAC 자리차지 표에서 vert offset이 있으면 문단 시작 y 전달
-        // layout_partial_table 내부에서 vert_offset을 적용하므로 이중 적용 방지
+        // 비-TAC 자리차지 표에서 vert offset이 있으면 문단 시작 y 전달.
+        // layout_partial_table 내부에서 vert_offset을 적용하므로 이중 적용 방지.
+        // [Task #712] HwpUnit=u32 이라 `vertical_offset > 0` 가드는 음수 비트표현
+        // (예: -1796 HU = 4294965500u32) 도 통과시킴. signed 비교로 정정.
         let pt_y_start = if let Some(para) = paragraphs.get(para_index) {
             if let Some(Control::Table(t)) = para.controls.get(control_index) {
                 if !t.common.treat_as_char
                     && matches!(t.common.text_wrap, crate::model::shape::TextWrap::TopAndBottom)
                     && matches!(t.common.vert_rel_to, crate::model::shape::VertRelTo::Para)
-                    && t.common.vertical_offset > 0
+                    && (t.common.vertical_offset as i32) > 0
                 {
                     para_start_y.get(&para_index).copied().unwrap_or(y_offset)
                 } else {
